+++ NEWS.md
@@ -5,6 +5,7 @@
 * Renamed `dictionary_sugar()` to `dictionary_sugar_get()`.
 * Added function `cite_bib()` and Rd macro `\cite{}` to insert entries from
   bibtex files into roxygen documentation.
+* `unnest()` now can creates list columns for non atomic items.
 
 # mlr3misc 0.1.4